refactor: 代码评审修复批次 — 数据源认证安全加固 + 架构规范化
🔴 严重修复:
- TwelveData/Tiingo API Key 从 URL 查询参数改为 Authorization Header
- schema.sql 从生产数据库导出覆盖，修复表结构不一致

🟡 改进:
- TigerOpenPythonBridge/MarketDataSourceRouterImpl/TiingoDataSourceStrategy 字段注入→构造函数注入
- 提取 WindowConstants 统一窗口常量
- 移除 ScanOrchestratorServiceImpl 无效 @Transactional
- ScreeningServiceImpl bars 判断改为动态最小窗口
- YFinanceStockServiceImpl @SuppressWarnings → TypeReference
- KLineIterator.setTime 异常日志/ResilientHttpExecutor 退避上限 300s/约束名修正

📋 新增:
- CacheProperties/GapFillProperties/SchedulingProperties
- LatestBarsCacheService 拆分
- E2E 测试 + 多项单元测试

🧹 清理:
- 删除 ScreeningPageNewController(已合并)/PatternEvaluateServiceImplTest(已迁移)
- .bak 文件移除, logs/ 加入 .gitignore

+++ src/main/java/com/stock/invest/model/StockInfo.java
@@ -14,7 +14,7 @@ public class StockInfo {
     private String name;
     private double currentPrice;
     private double openPrice;
-    private int volume;
+    private long volume;
     private double change;
     private double changePercent;
 }
